+++ src/proxy/proxy.h
@@ -2,8 +2,9 @@
 #define PROXY_H
 
 #include <stddef.h>
+#include "Clay/clay.h"
 
-int proxy_start(int port);
+int proxy_start(uint16 port);
 void proxy_shutdown(void);
 
 #endif
